fix: use in-cluster config for vcluster operations
Use in-cluster connection type instead of local when installing/uninstalling
vcluster for managed workspaces. The local type tries to read ~/.kube/config
which doesn't exist when the controller is running as a pod in the cluster.

Fixes error: failed to load local kubeconfig from /home/nonroot/.kube/config:
no such file or directory